requirements.txt

chromadb
langchain
langchain-community==0.2.10
langchain_experimental
langchainhub
llama-cpp-python
sentence-transformers
tiktoken
umap-learn
matplotlib
bs4